fixed a bug when mapmanager spawned incorrectly, fix a bug when dropped gem didnt blow away by explosion force on client

+++ Assets/Scripts/Map/MapManager.cs
@@ -381,7 +381,7 @@ public class MapManager : NetworkBehaviour, IMapManager
     public Vector2Int GetMapSize() => mapSize;
 
     public Vector2Int PositionToIndex(Vector2 position) => 
-    new Vector2Int(Mathf.FloorToInt(position.x - rootX), Mathf.FloorToInt(position.y - rootY));
+    new Vector2Int(Mathf.FloorToInt(position.x - (float)rootX), Mathf.FloorToInt(position.y - rootY));
 
     [Server]
     public bool TrySpawnDiggableAtIndex(Vector2Int idx, DiggableType diggable, GameObject prefab)
